+++ Libraries/LibC/sys/types.h
@@ -56,7 +56,7 @@ typedef uint16_t mode_t;
 typedef uint32_t nlink_t;
 typedef uint32_t blksize_t;
 typedef uint32_t blkcnt_t;
-typedef uint32_t time_t;
+typedef int64_t time_t;
 typedef uint32_t useconds_t;
 typedef int32_t suseconds_t;
 typedef uint32_t clock_t;
